Fix: Разрешить конфликт "use client" и `generateMetadata` в RootLayout
Apply AI assistant changes to 2 files.
Ref: Fix: Разрешить конфликт "use client" и `generateMetadata` в RootLayout...
